İşlemin başarılı olup olmadığı kontrol edildi. Gelen cevaba göre başarılı veya başarısız toast'ı eklendi.
